continue model download
wget -c makes it possible to continue a download after a broken connection or a stopped wget. The &&-operator tells only do the next command if the first command was successful. So decompressing the file is only done if download was complete and remove is done if decompress was successful what means the download was also successful.